Add union.cpp and linkList.h
This commit adds the files union.cpp and linkList.h to the repository. The union.cpp file contains code for creating and manipulating a sequence list, while the linkList.h file contains the implementation of a linked list data structure. These files are necessary for future development and will be used in upcoming features.